requirements.txt
requests~=2.32.3
python-dotenv~=1.1.0
openai~=1.77.0
streamlit~=1.45.0

pygame~=2.6.1
pandas~=2.2.3
tqdm~=4.67.1
numpy~=2.2.5
faiss-cpu~=1.11.0
rank-bm25~=0.2.2
